moving 2D sensor FoV calculation into data model out of visual asset
